# [storage] Optimize `journal::replay()`