--- a/consensus/src/ordered_broadcast/config.rs
+++ b/consensus/src/ordered_broadcast/config.rs
@@ -76,6 +76,9 @@ pub struct Config<
     /// Upon replaying a journal, the number of entries to replay concurrently.
     pub journal_replay_concurrency: usize,
 
+    /// The number of bytes to look ahead when replaying a journal.
+    pub journal_replay_lookahead: usize,
+
     /// Compression level for the journal.
     pub journal_compression: Option<u8>,
 }