--- a/consensus/src/ordered_broadcast/engine.rs
+++ b/consensus/src/ordered_broadcast/engine.rs
@@ -141,7 +141,7 @@ pub struct Engine<
     journal_replay_concurrency: usize,
 
     // The number of bytes to look ahead when replaying a journal.
-    journal_replay_lookahead: usize,
+    journal_replay_buffer: usize,
 
     // A prefix for the journal names.
     // The rest of the name is the hex-encoded public keys of the relevant sequencer.
@@ -234,7 +234,7 @@ impl<
             pending_verifies: FuturesPool::default(),
             journal_heights_per_section: cfg.journal_heights_per_section,
             journal_replay_concurrency: cfg.journal_replay_concurrency,
-            journal_replay_lookahead: cfg.journal_replay_lookahead,
+            journal_replay_buffer: cfg.journal_replay_buffer,
             journal_name_prefix: cfg.journal_name_prefix,
             journal_compression: cfg.journal_compression,
             journals: BTreeMap::new(),
@@ -974,10 +974,7 @@ impl<
 
             // Prepare the stream
             let stream = journal
-                .replay(
-                    self.journal_replay_concurrency,
-                    self.journal_replay_lookahead,
-                )
+                .replay(self.journal_replay_concurrency, self.journal_replay_buffer)
                 .await
                 .expect("unable to replay journal");
             pin_mut!(stream);