--- a/consensus/src/ordered_broadcast/mod.rs
+++ b/consensus/src/ordered_broadcast/mod.rs
@@ -234,7 +234,7 @@ mod tests {
                     priority_proposals: false,
                     journal_heights_per_section: 10,
                     journal_replay_concurrency: 1,
-                    journal_replay_lookahead: 4096,
+                    journal_replay_buffer: 4096,
                     journal_name_prefix: format!("ordered-broadcast-seq/{}/", validator),
                     journal_compression: Some(3),
                 },
@@ -830,7 +830,7 @@ mod tests {
                         priority_proposals: false,
                         journal_heights_per_section: 10,
                         journal_replay_concurrency: 1,
-                        journal_replay_lookahead: 4096,
+                        journal_replay_buffer: 4096,
                         journal_name_prefix: format!("ordered-broadcast-seq/{}/", validator),
                         journal_compression: Some(3),
                     },
@@ -879,7 +879,7 @@ mod tests {
                         priority_proposals: false,
                         journal_heights_per_section: 10,
                         journal_replay_concurrency: 1,
-                        journal_replay_lookahead: 4096,
+                        journal_replay_buffer: 4096,
                         journal_name_prefix: format!(
                             "ordered-broadcast-seq/{}/",
                             sequencer.public_key()