--- a/consensus/src/simplex/actors/voter/mod.rs
+++ b/consensus/src/simplex/actors/voter/mod.rs
@@ -34,6 +34,7 @@ pub struct Config<
     pub activity_timeout: View,
     pub skip_timeout: View,
     pub replay_concurrency: usize,
+    pub replay_lookahead: usize,
 }
 
 #[cfg(test)]
@@ -124,6 +125,7 @@ mod tests {
                 activity_timeout: 10,
                 skip_timeout: 10,
                 replay_concurrency: 1,
+                replay_lookahead: 1024 * 1024,
             };
             let (actor, mut mailbox) = Actor::new(context.clone(), cfg);