--- a/consensus/src/simplex/config.rs
+++ b/consensus/src/simplex/config.rs
@@ -45,7 +45,7 @@ pub struct Config<
     pub replay_concurrency: usize,
 
     /// Number of bytes to look ahead when replaying during startup.
-    pub replay_lookahead: usize,
+    pub replay_buffer: usize,
 
     /// Amount of time to wait for a leader to propose a payload
     /// in a view.