--- a/consensus/src/simplex/engine.rs
+++ b/consensus/src/simplex/engine.rs
@@ -65,7 +65,7 @@ impl<
                 activity_timeout: cfg.activity_timeout,
                 skip_timeout: cfg.skip_timeout,
                 replay_concurrency: cfg.replay_concurrency,
-                replay_lookahead: cfg.replay_lookahead,
+                replay_buffer: cfg.replay_buffer,
             },
         );