--- a/consensus/src/simplex/mod.rs
+++ b/consensus/src/simplex/mod.rs
@@ -324,6 +324,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
+                    replay_lookahead: 1024 * 1024,
                 };
                 let engine = Engine::new(context.with_label("engine"), cfg);
                 let (voter, resolver) = registrations
@@ -567,6 +568,7 @@ mod tests {
                         fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                         fetch_concurrent: 1,
                         replay_concurrency: 1,
+                        replay_lookahead: 1024 * 1024,
                     };
                     let engine = Engine::new(context.with_label("engine"), cfg);
                     let (voter_network, resolver_network) = registrations
@@ -734,6 +736,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
+                    replay_lookahead: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)
@@ -851,6 +854,7 @@ mod tests {
                 fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                 fetch_concurrent: 1,
                 replay_concurrency: 1,
+                replay_lookahead: 1024 * 1024,
             };
             let (voter, resolver) = registrations
                 .remove(&validator)
@@ -971,6 +975,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
+                    replay_lookahead: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)
@@ -1165,6 +1170,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
+                    replay_lookahead: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)
@@ -1321,6 +1327,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
+                    replay_lookahead: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)
@@ -1470,6 +1477,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
+                    replay_lookahead: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)
@@ -1647,6 +1655,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
+                    replay_lookahead: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)
@@ -1812,6 +1821,7 @@ mod tests {
                         fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                         fetch_concurrent: 1,
                         replay_concurrency: 1,
+                        replay_lookahead: 1024 * 1024,
                     };
                     let (voter, resolver) = registrations
                         .remove(&validator)
@@ -1978,6 +1988,7 @@ mod tests {
                         fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                         fetch_concurrent: 1,
                         replay_concurrency: 1,
+                        replay_lookahead: 1024 * 1024,
                     };
                     let (voter, resolver) = registrations
                         .remove(&validator)
@@ -2140,6 +2151,7 @@ mod tests {
                         fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                         fetch_concurrent: 1,
                         replay_concurrency: 1,
+                        replay_lookahead: 1024 * 1024,
                     };
                     let (voter, resolver) = registrations
                         .remove(&validator)
@@ -2274,6 +2286,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
+                    replay_lookahead: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)