--- a/consensus/src/simplex/mod.rs
+++ b/consensus/src/simplex/mod.rs
@@ -324,7 +324,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
-                    replay_lookahead: 1024 * 1024,
+                    replay_buffer: 1024 * 1024,
                 };
                 let engine = Engine::new(context.with_label("engine"), cfg);
                 let (voter, resolver) = registrations
@@ -568,7 +568,7 @@ mod tests {
                         fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                         fetch_concurrent: 1,
                         replay_concurrency: 1,
-                        replay_lookahead: 1024 * 1024,
+                        replay_buffer: 1024 * 1024,
                     };
                     let engine = Engine::new(context.with_label("engine"), cfg);
                     let (voter_network, resolver_network) = registrations
@@ -736,7 +736,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
-                    replay_lookahead: 1024 * 1024,
+                    replay_buffer: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)
@@ -854,7 +854,7 @@ mod tests {
                 fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                 fetch_concurrent: 1,
                 replay_concurrency: 1,
-                replay_lookahead: 1024 * 1024,
+                replay_buffer: 1024 * 1024,
             };
             let (voter, resolver) = registrations
                 .remove(&validator)
@@ -975,7 +975,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
-                    replay_lookahead: 1024 * 1024,
+                    replay_buffer: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)
@@ -1170,7 +1170,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
-                    replay_lookahead: 1024 * 1024,
+                    replay_buffer: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)
@@ -1327,7 +1327,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
-                    replay_lookahead: 1024 * 1024,
+                    replay_buffer: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)
@@ -1477,7 +1477,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
-                    replay_lookahead: 1024 * 1024,
+                    replay_buffer: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)
@@ -1655,7 +1655,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
-                    replay_lookahead: 1024 * 1024,
+                    replay_buffer: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)
@@ -1821,7 +1821,7 @@ mod tests {
                         fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                         fetch_concurrent: 1,
                         replay_concurrency: 1,
-                        replay_lookahead: 1024 * 1024,
+                        replay_buffer: 1024 * 1024,
                     };
                     let (voter, resolver) = registrations
                         .remove(&validator)
@@ -1988,7 +1988,7 @@ mod tests {
                         fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                         fetch_concurrent: 1,
                         replay_concurrency: 1,
-                        replay_lookahead: 1024 * 1024,
+                        replay_buffer: 1024 * 1024,
                     };
                     let (voter, resolver) = registrations
                         .remove(&validator)
@@ -2151,7 +2151,7 @@ mod tests {
                         fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                         fetch_concurrent: 1,
                         replay_concurrency: 1,
-                        replay_lookahead: 1024 * 1024,
+                        replay_buffer: 1024 * 1024,
                     };
                     let (voter, resolver) = registrations
                         .remove(&validator)
@@ -2286,7 +2286,7 @@ mod tests {
                     fetch_rate_per_peer: Quota::per_second(NonZeroU32::new(1).unwrap()),
                     fetch_concurrent: 1,
                     replay_concurrency: 1,
-                    replay_lookahead: 1024 * 1024,
+                    replay_buffer: 1024 * 1024,
                 };
                 let (voter, resolver) = registrations
                     .remove(&validator)