--- a/consensus/src/threshold_simplex/actors/voter/actor.rs
+++ b/consensus/src/threshold_simplex/actors/voter/actor.rs
@@ -514,6 +514,7 @@ pub struct Actor<
     partition: String,
     compression: Option<u8>,
     replay_concurrency: usize,
+    replay_lookahead: usize,
     journal: Option<Journal<E, Voter<D>>>,
 
     genesis: Option<D>,
@@ -610,6 +611,7 @@ impl<
                 partition: cfg.partition,
                 compression: cfg.compression,
                 replay_concurrency: cfg.replay_concurrency,
+                replay_lookahead: cfg.replay_lookahead,
                 journal: None,
 
                 genesis: None,
@@ -1799,7 +1801,7 @@ impl<
         self.enter_view(1, group::Signature::zero());
 
         // Initialize journal
-        let mut journal = Journal::<_, Voter<D>>::init(
+        let journal = Journal::<_, Voter<D>>::init(
             self.context.with_label("journal"),
             JConfig {
                 partition: self.partition.clone(),
@@ -1813,7 +1815,7 @@ impl<
         // Rebuild from journal
         {
             let stream = journal
-                .replay(self.replay_concurrency)
+                .replay(self.replay_concurrency, self.replay_lookahead)
                 .await
                 .expect("unable to replay journal");
             pin_mut!(stream);