--- a/consensus/src/threshold_simplex/config.rs
+++ b/consensus/src/threshold_simplex/config.rs
@@ -44,6 +44,9 @@ pub struct Config<
     /// Number of views to replay concurrently during startup.
     pub replay_concurrency: usize,
 
+    /// Number of bytes to look ahead when replaying during startup.
+    pub replay_lookahead: usize,
+
     /// Amount of time to wait for a leader to propose a payload
     /// in a view.
     pub leader_timeout: Duration,