--- a/consensus/src/threshold_simplex/engine.rs
+++ b/consensus/src/threshold_simplex/engine.rs
@@ -79,6 +79,7 @@ impl<
                 activity_timeout: cfg.activity_timeout,
                 skip_timeout: cfg.skip_timeout,
                 replay_concurrency: cfg.replay_concurrency,
+                replay_lookahead: cfg.replay_lookahead,
             },
         );