--- a/runtime/src/lib.rs
+++ b/runtime/src/lib.rs
@@ -38,9 +38,7 @@ mod network;
 mod storage;
 pub mod telemetry;
 mod utils;
-pub use utils::{
-    create_pool, reschedule, Handle, RwLock, RwLockReadGuard, RwLockWriteGuard, Signal, Signaler,
-};
+pub use utils::*;
 
 /// Prefix for runtime metrics.
 const METRICS_PREFIX: &str = "runtime";