Resolves: #897 
Future Work: #902

## Results

_Reproduce by running the command (from inside `storage`): `cargo bench --bench journal -- journal::fixed_replay`._

### `main`
```
journal::fixed_replay/items=1000 size=32
                        time:   [11.362 ms 11.393 ms 11.428 ms]
journal::fixed_replay/items=10000 size=32
                        time:   [111.57 ms 112.68 ms 113.72 ms]
journal::fixed_replay/items=100000 size=32
                        time:   [1.1191 s 1.1302 s 1.1427 s]
journal::fixed_replay/items=500000 size=32
                        time:   [5.5083 s 5.5745 s 5.6678 s]
```

### `diff` (-99%)
```
journal::fixed_replay/items=1000 buffer=128 size=32
                        time:   [3.2873 ms 3.2919 ms 3.2979 ms]
journal::fixed_replay/items=1000 buffer=16384 size=32
                        time:   [155.12 µs 155.63 µs 156.25 µs]
journal::fixed_replay/items=1000 buffer=65536 size=32
                        time:   [123.50 µs 123.82 µs 124.03 µs]
journal::fixed_replay/items=1000 buffer=1048576 size=32
                        time:   [127.67 µs 128.02 µs 128.49 µs]
journal::fixed_replay/items=10000 buffer=128 size=32
                        time:   [32.720 ms 32.818 ms 32.918 ms]
journal::fixed_replay/items=10000 buffer=16384 size=32
                        time:   [1.4659 ms 1.4760 ms 1.4895 ms]
journal::fixed_replay/items=10000 buffer=65536 size=32
                        time:   [1.1765 ms 1.1800 ms 1.1862 ms]
journal::fixed_replay/items=10000 buffer=1048576 size=32
                        time:   [1.0822 ms 1.0890 ms 1.0998 ms]
journal::fixed_replay/items=100000 buffer=128 size=32
                        time:   [327.79 ms 329.47 ms 331.07 ms]
journal::fixed_replay/items=100000 buffer=16384 size=32
                        time:   [14.638 ms 14.674 ms 14.745 ms]
journal::fixed_replay/items=100000 buffer=65536 size=32
                        time:   [11.550 ms 11.584 ms 11.619 ms]
journal::fixed_replay/items=100000 buffer=1048576 size=32
                        time:   [10.691 ms 10.783 ms 10.975 ms]
journal::fixed_replay/items=500000 buffer=128 size=32
                        time:   [1.6123 s 1.6352 s 1.6552 s]
journal::fixed_replay/items=500000 buffer=16384 size=32
                        time:   [73.261 ms 73.406 ms 73.555 ms]
journal::fixed_replay/items=500000 buffer=65536 size=32
                        time:   [58.329 ms 58.443 ms 58.601 ms]
journal::fixed_replay/items=500000 buffer=1048576 size=32
                        time:   [55.060 ms 56.339 ms 57.447 ms]
```